Organize data360 specific code
